Photok-RecWare-Windows/0.2/requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
humanize==4.12.3
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pefile==2023.2.7
pycryptodome==3.23.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-magic-bin==0.4.14
pywin32-ctypes==0.2.3
qt-material==2.17
setuptools==80.9.0
